rename text to label
