+++ packages/@dcl/react-ecs/src/components/Label/types.ts
