+++ packages/@dcl/react-ecs/src/components/uiTransform/index.ts
@@ -13,37 +13,6 @@ import { PBUiTransform } from '@dcl/ecs/dist/components'
 
 export const CANVAS_ROOT_ENTITY = 0
 
-/**
- * @public
- */
-export function parseUiTransform(props: UiTransformProps = {}): PBUiTransform {
-  const {
-    position,
-    padding,
-    margin,
-    height,
-    minHeight,
-    maxHeight,
-    width,
-    maxWidth,
-    minWidth,
-    ...otherProps
-  } = props
-  return {
-    ...defaultUiTransform,
-    ...otherProps,
-    ...parsePosition(position, 'position'),
-    ...parsePosition(margin, 'margin'),
-    ...parsePosition(padding, 'padding'),
-    ...parseSize(height, 'height'),
-    ...parseSize(minHeight, 'minHeight'),
-    ...parseSize(maxHeight, 'maxHeight'),
-    ...parseSize(width, 'width'),
-    ...parseSize(minWidth, 'minWidth'),
-    ...parseSize(maxWidth, 'maxWidth')
-  }
-}
-
 const defaultUiTransform: PBUiTransform = {
   parent: CANVAS_ROOT_ENTITY,
   rightOf: 0,
@@ -92,4 +61,36 @@ const defaultUiTransform: PBUiTransform = {
   flexBasisUnit: YGUnit.YGU_UNDEFINED,
   widthUnit: YGUnit.YGU_UNDEFINED,
   heightUnit: YGUnit.YGU_UNDEFINED
+}
+
+/**
+ * @public
+ */
+/*#__PURE__*/
+export function parseUiTransform(props: UiTransformProps = {}): PBUiTransform {
+  const {
+    position,
+    padding,
+    margin,
+    height,
+    minHeight,
+    maxHeight,
+    width,
+    maxWidth,
+    minWidth,
+    ...otherProps
+  } = props
+  return {
+    ...defaultUiTransform,
+    ...otherProps,
+    ...parsePosition(position, 'position'),
+    ...parsePosition(margin, 'margin'),
+    ...parsePosition(padding, 'padding'),
+    ...parseSize(height, 'height'),
+    ...parseSize(minHeight, 'minHeight'),
+    ...parseSize(maxHeight, 'maxHeight'),
+    ...parseSize(width, 'width'),
+    ...parseSize(minWidth, 'minWidth'),
+    ...parseSize(maxWidth, 'maxWidth')
+  }
 }
